[#IC-189] refactor exemption certificate tests
